pyproject.toml
[project]
name = "va-rag"
version = "0.1.0"
description = "Visually Augmented RAG stack for SharePoint manuals (ColPali + LlamaIndex + DSPy + PydanticAI)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "msgraph-sdk>=1.5.0",
  "azure-identity>=1.16.0",
  "qdrant-client>=1.9.0",
  "byaldi>=0.0.4",
  "llama-index>=0.10.0",
  "llama-index-core>=0.10.0",
  "llama-index-readers-microsoft-sharepoint>=0.1.3",
  "llama-index-embeddings-openai>=0.1.10",
  "llama-index-vector-stores-qdrant>=0.1.4",
  "pydantic>=2.7.0",
  "pydantic-ai>=0.0.7",
  "docling>=0.7.0",
  "dspy-ai>=2.4.0",
  "pillow>=10.0.0",
  "pypdf>=4.0.0",
  "pypdfium2>=4.30.0",
  "python-dotenv>=1.0.1",
  "loguru>=0.7.2",
  "requests>=2.32.0",
  "llama-index-embeddings-huggingface>=0.6.1",
  "sentence-transformers>=5.2.0",
  "azure-storage-blob>=12.19.0",
  "docling-core>=0.7.0",
]

[tool.uv]
dev-dependencies = [
  "pytest>=8.0.0",
  "pytest-mock>=3.12.0",
  "types-requests"
]

[build-system]
requires = ["setuptools>=68.0.0"]
build-backend = "setuptools.build_meta"

